added which user adds the product and only the user that added the product may view and edit the details of the product

+++ Controllers/ProductController.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
@@ -13,6 +14,8 @@ namespace TamansShop.Controllers
     public class ProductController : Controller
     {
         private readonly TamansShopContext _context;
+        private readonly ILogger<ProductController> _logger;
+
 
         public ProductController(TamansShopContext context)
         {
@@ -22,15 +25,14 @@ namespace TamansShop.Controllers
         // GET: Product
         public async Task<IActionResult> Index()
         {
-              return _context.Products != null ? 
-                          View(await _context.Products.ToListAsync()) :
-                          Problem("Entity set 'TamansShopContext.Products'  is null.");
+            var products = await _context.Products.ToListAsync();
+            return View(products);
         }
 
         // GET: Product/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null || _context.Products == null)
+            if (id == null)
             {
                 return NotFound();
             }
@@ -52,21 +54,17 @@ namespace TamansShop.Controllers
         }
 
         // POST: Product/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Title,Description,Price,ImagePath")] Product product)
         {
-            if (ModelState.IsValid)
-            {
+            string currentUserId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            product.UserId = currentUserId;
                 _context.Add(product);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
-            }
-            return View(product);
+            
         }
-
         // GET: Product/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -79,6 +77,11 @@ namespace TamansShop.Controllers
             if (product == null)
             {
                 return NotFound();
+            }
+            // Check if the current user is the owner of the product
+            if (product.UserId != User.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+            {
+                return Forbid(); // Return a forbidden status
             }
             return View(product);
         }
@@ -93,6 +96,11 @@ namespace TamansShop.Controllers
             if (id != product.Id)
             {
                 return NotFound();
+            }
+            // Check if the current user is the owner of the product
+            if (product.UserId != User.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+            {
+                return Forbid(); // Return a forbidden status
             }
 
             if (ModelState.IsValid)
@@ -132,6 +140,11 @@ namespace TamansShop.Controllers
             {
                 return NotFound();
             }
+            // Check if the current user is the owner of the product
+            if (product.UserId != User.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+            {
+                return Forbid(); // Return a forbidden status
+            }
 
             return View(product);
         }
@@ -146,6 +159,11 @@ namespace TamansShop.Controllers
                 return Problem("Entity set 'TamansShopContext.Products'  is null.");
             }
             var product = await _context.Products.FindAsync(id);
+            // Check if the current user is the owner of the product
+            if (product.UserId != User.FindFirst(ClaimTypes.NameIdentifier)?.Value)
+            {
+                return Forbid(); // Return a forbidden status
+            }
             if (product != null)
             {
                 _context.Products.Remove(product);
